pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mobo-kit"
version = "0.1.0"
description = "Multi-objective Bayesian optimization toolkit for functional thin-film fabrication"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Ethan Schwartz", email = "ebuddy23@uw.edu"},
    {name = "Daniel Abdoue", email = "dabdoue@ucsd.edu"},
    {name = "Nicky Evans", email = "nickye17@mit.edu"},
    {name = "Tonio Buonassisi", email = "buonassi@mit.edu"},
]
maintainers = [
    {name = "Ethan Schwartz", email = "ebuddy23@uw.edu"},
]
keywords = [
    "bayesian-optimization",
    "multi-objective",
    "materials-science", 
    "thin-films",
    "machine-learning",
    "optimization",
    "gaussian-processes"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Topic :: Scientific/Engineering :: Physics",
]
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "pyyaml>=6.0",
    "scipy>=1.7.0",
    "torch>=1.12.0",
    "gpytorch>=1.8.0",
    "botorch>=0.8.0",
    "emukit>=0.4.10",
    "shap>=0.41.0",
    "pyDOE>=0.3.8",
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=22.0",
    "flake8>=4.0",
    "mypy>=0.950",
]
jupyter = [
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
    "notebook>=6.0.0",
]
all = [
    "mobo-fom[dev,jupyter]",
]

[project.urls]
Homepage = "https://github.com/PV-Lab/MOBO-FOM"
Repository = "https://github.com/PV-Lab/MOBO-FOM"
Documentation = "https://github.com/PV-Lab/MOBO-FOM#readme"
"Bug Tracker" = "https://github.com/PV-Lab/MOBO-FOM/issues"

[project.scripts]
mobo-kit = "mobo_kit.cli:main"
mobo-kit-run = "mobo_kit.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
